Update main.py - IMPORTANT
Remove time resync functionality and adjusted startup procedure. Wireless chip will only be active for as short a time as possible and will only be active BEFORE any servo movement is performed. 

Testing has revealed that enabling the wireless chip at any time when servos may also be active can cause too high of a power draw resulting in either the system rebooting or possibly overheating/causing a fault in the power circuitry possibly resulting in damage.

It is highly advised to NOT enable the wifi chip during operation of the clock. It appears to be fine during initial startup, but should be disabled otherwise. Code has been updated to avoid these issues from occuring.